Create View Structure
In this application will have two views groups: One for the normal user (reader) and other for the admin (writer).
In this way, all the views related to the user will be stored in the blog folder and the ones for the admin in the admin folder.
There will be two layouts (one for the Reader and other for the Writer)